feat(GameBoard): implement part duplication logic and update onCopyPart handler
