+++ frontend/src/features/prototype/components/molecules/PartPropertyMenu.tsx
@@ -17,16 +17,15 @@ import {
 import { IoMdMove } from 'react-icons/io';
 
 import { useImages } from '@/api/hooks/useImages';
-import { Part, PartProperty } from '@/api/types';
+import { Part } from '@/api/types';
 import NumberInput from '@/components/atoms/NumberInput';
 import TextInput from '@/components/atoms/TextInput';
 import ColorPicker from '@/features/prototype/components/atoms/ColorPicker';
 import PartPropertyMenuButton from '@/features/prototype/components/atoms/PartPropertyMenuButton';
-import { COLORS, GAME_BOARD_SIZE } from '@/features/prototype/constants';
+import { COLORS } from '@/features/prototype/constants';
 import useDraggablePartPropertyMenu from '@/features/prototype/hooks/useDraggablePartPropertyMenu';
 import { usePartReducer } from '@/features/prototype/hooks/usePartReducer';
 import {
-  AddPartProps,
   DeleteImageProps,
   PartPropertyUpdate,
   PartPropertyWithImage,
@@ -41,8 +40,6 @@ export type PartPropertyMenuProps = {
   parts: Part[];
   // パーツのプロパティ
   properties: PartPropertyWithImage[];
-  // パーツを追加時の処理
-  onAddPart: ({ part, properties }: AddPartProps) => void;
   // パーツを削除時の処理
   onDeletePart: () => void;
   // 画像をクリア時の処理
@@ -53,15 +50,17 @@ export type PartPropertyMenuProps = {
     side,
     emitUpdate,
   }: DeleteImageProps) => void;
+  // パーツ複製時の処理
+  onCopyPart: () => void;
 };
 
 export default function PartPropertyMenu({
   selectedPartIds,
   parts,
   properties,
-  onAddPart,
   onDeletePart,
   onDeleteImage,
+  onCopyPart,
 }: PartPropertyMenuProps) {
   const selectedPartId = selectedPartIds[0];
   const showMenu = selectedPartIds.length === 1;
@@ -112,89 +111,7 @@ export default function PartPropertyMenu({
 
   const { containerRef, position, isDragging, handleDragStart } =
     useDraggablePartPropertyMenu();
-  /**
-   * パーツを複製する
-   */
-  const handleCopyPart = () => {
-    // 選択中のパーツが存在しない、またはプロパティが存在しない場合
-    if (!selectedPart || !selectedPartProperties) return;
-
-    // 新しいパーツ
-    /**
-     * 複製パーツの配置位置を決定する
-     * - 既存パーツの右→下→左→上の順で試し、ボード外なら次候補へ
-     * - 全て収まらない場合は右候補を基準にボード内へクランプ
-     * @param part 対象パーツ
-     * @returns {{ x: number; y: number }} 配置位置
-     */
-    const computeCopyPosition = (part: Part): { x: number; y: number } => {
-      const boardMaxX = GAME_BOARD_SIZE - part.width;
-      const boardMaxY = GAME_BOARD_SIZE - part.height;
-
-      const positionCandidates = [
-        { x: part.position.x + part.width, y: part.position.y },
-        { x: part.position.x, y: part.position.y + part.height },
-        { x: part.position.x - part.width, y: part.position.y },
-        { x: part.position.x, y: part.position.y - part.height },
-      ];
-
-      const fit = positionCandidates.find(
-        (c) => c.x >= 0 && c.y >= 0 && c.x <= boardMaxX && c.y <= boardMaxY
-      );
-      if (fit) return fit;
-
-      // どれも収まらない場合はクランプしてボード内に配置
-      const clamped = positionCandidates[0];
-      return {
-        x: Math.min(Math.max(0, clamped.x), boardMaxX),
-        y: Math.min(Math.max(0, clamped.y), boardMaxY),
-      };
-    };
-
-    const newPart: Omit<
-      Part,
-      'id' | 'prototypeId' | 'order' | 'createdAt' | 'updatedAt'
-    > = {
-      type: selectedPart.type,
-      position: computeCopyPosition(selectedPart),
-      width: selectedPart.width,
-      height: selectedPart.height,
-    };
-
-    // カードパーツの場合
-    if (selectedPart.type === 'card') {
-      newPart.frontSide = selectedPart.frontSide;
-    } else {
-      // カード以外のパーツの場合はデフォルトで'front'を設定
-      newPart.frontSide = 'front';
-    }
-
-    // 手札パーツの場合
-    if (selectedPart.type === 'hand') {
-      newPart.ownerId = selectedPart.ownerId;
-    }
-
-    // 新しいパーツのプロパティ
-    const newPartProperties: Omit<
-      PartProperty,
-      'id' | 'partId' | 'createdAt' | 'updatedAt'
-    >[] = selectedPartProperties
-      .filter(({ side }) => {
-        // カードパーツの場合は両面、それ以外は'front'のみ
-        return selectedPart.type === 'card' ? true : side === 'front';
-      })
-      .map(({ side, name, description, color, imageId, textColor }) => {
-        return {
-          side,
-          name,
-          description,
-          color,
-          textColor,
-          imageId,
-        };
-      });
-    onAddPart({ part: newPart, properties: newPartProperties });
-  };
+  // ...existing code... (copy logic moved to GameBoard)
 
   /**
    * プロパティの値が変化している場合のみ更新処理を呼ぶ
@@ -345,9 +262,7 @@ export default function PartPropertyMenu({
               <PartPropertyMenuButton
                 text="複製"
                 icon={<FaRegCopy className="h-3 w-3" />}
-                onClick={() => {
-                  handleCopyPart();
-                }}
+                onClick={() => onCopyPart()}
               />
               <PartPropertyMenuButton
                 text="削除"
